+++ Presentation/CMdm.UI.Web/Helpers/CrossCutting/Security/CustomMembershipProvider.cs
@@ -144,7 +144,7 @@ namespace CMdm.UI.Web.Helpers.CrossCutting.Security
             ICollection<CM_USER_ROLE_XREF> UserRoleColl = new Collection<CM_USER_ROLE_XREF>();
             CM_USER_PROFILE up = new CM_USER_PROFILE();
             string passwordHash = pwdManager.GeneratePasswordHash(password, out salt);
-            decimal profileId = 0;
+            int profileId = 0;
             using (var cdma = new AppDbContext())
             {
                 var usr = new CM_USER_PROFILE
